extracting the indicator coordinates calculating
